+++ R/map.R
@@ -11,7 +11,6 @@
 #'
 #'   The `_if` and `_at` variants take a predicate function `.p` that
 #'   determines which elements of `.x` are transformed with `.f`.
-#'   transform.
 #'
 #' * `map_lgl()`, `map_int()`, `map_dbl()` and `map_chr()` return
 #'   vectors of the corresponding type (or die trying).
@@ -30,9 +29,9 @@
 #'   objects, a string indicating the name of a logical element in the
 #'   inner lists. Only those elements where `.p` evaluates to
 #'   `TRUE` will be modified.
-#' @param .at A character vector of names or a numeric vector of
-#'   positions. Only those elements corresponding to `.at` will be
-#'   modified.
+#' @param .at A character vector of names, positive numeric vector of
+#'   positions to include, or a negative numeric vector of positions to
+#'   exlude. Only those elements corresponding to `.at` will be modified.
 #' @param ... Additional arguments passed on to `.f`.
 #' @return All functions return a vector the same length as `.x`.
 #'